YH-1378: update tabs
