+++ src/pages/admin/ResourcesPage/index.ts
@@ -1,5 +1,5 @@
-export { default as ResourcesAdminPage } from './ui/ResourcesPage';
-export { ResourcesRequestsTab, resourcesRequestsTabReducer } from './ui/tabs/ResourcesRequestsTab';
-export type { ResourcesRequestsTabState } from './ui/tabs/ResourcesRequestsTab';
-export { ResourcesAllTabAsync, resourcesAllTabReducer } from './ui/tabs/ResourcesAllTab';
-export type { ResourcesAllTabState } from './ui/tabs/ResourcesAllTab';
+export { ResourcesPage } from './ui/ResourcesPage.lazy';
+export type { ResourcesRequestsTabState } from './model/types/resourcesRequestsTabTypes';
+export type { ResourcesAllTabState } from './model/types/resourcesAllTabTypes';
+export { resourcesAllTabReducer } from './model/slice/resourcesAllTabSlice';
+export { resourcesRequestsTabReducer } from './model/slice/resourcesRequestsTabSlice';
