+++ src/pages/admin/ResourcesPage/model/hooks/useGetResourcesPageTabs.tsx
@@ -1,0 +1,29 @@
+import { useTranslation } from 'react-i18next';
+
+import { i18Namespace } from '@/shared/config/i18n';
+import { ResourceRequests } from '@/shared/config/i18n/i18nTranslations';
+import { Tab } from '@/shared/ui/Tabs';
+
+import { ResourcesAllTab } from '../../ui/tabs/ResourcesAllTab/ResourcesAllTab/ResourcesAllTab';
+import { ResourcesRequestsTab } from '../../ui/tabs/ResourcesRequestsTab/ResourcesRequestsTab/ResourcesRequestsTab';
+
+type AdminResourcesTabId = 'all' | 'requests';
+
+export const useGetResourcesPageTabs = () => {
+	const { t } = useTranslation(i18Namespace.resources);
+
+	const tabs: Tab<AdminResourcesTabId>[] = [
+		{
+			id: 'all',
+			label: t(ResourceRequests.TABS_ALL),
+			Component: ResourcesAllTab,
+		},
+		{
+			id: 'requests',
+			label: t(ResourceRequests.TABS_REQUESTS),
+			Component: ResourcesRequestsTab,
+		},
+	];
+
+	return tabs;
+};
