+++ src/pages/admin/ResourcesPage/model/selectors/resourcesAllTabSelectors.ts
