+++ src/pages/admin/ResourcesPage/model/slice/resourcesAllTabSlice.ts
