+++ src/pages/admin/ResourcesPage/model/types/resourcesAllTabTypes.ts
