+++ src/pages/admin/ResourcesPage/ui/ResourcesPage.lazy.tsx
@@ -1,0 +1,3 @@
+import { lazy } from 'react';
+
+export const ResourcesPage = lazy(() => import('./ResourcesPage'));
