+++ src/pages/admin/ResourcesPage/ui/ResourcesPage.tsx
@@ -1,37 +1,10 @@
-import { useMemo } from 'react';
-import { useTranslation } from 'react-i18next';
-
-import { i18Namespace } from '@/shared/config/i18n';
-import { ResourceRequests } from '@/shared/config/i18n/i18nTranslations';
 import { Flex } from '@/shared/ui/Flex';
-import { Tab, Tabs, useTabs } from '@/shared/ui/Tabs';
+import { Tabs, useTabs } from '@/shared/ui/Tabs';
 
-import ResourcesAllTab from './tabs/ResourcesAllTab/ui/ResourcesAllTab/ResourcesAllTab';
-import { ResourcesRequestsTab } from './tabs/ResourcesRequestsTab';
+import { useGetResourcesPageTabs } from '../model/hooks/useGetResourcesPageTabs';
 
-/**
- * Page showing info about all the created questions
- * @constructor
- */
-
-type AdminResourcesTabId = 'all' | 'requests';
-
-const createAdminResourcesTabs = (t: (key: string) => string): Tab<AdminResourcesTabId>[] => [
-	{
-		id: 'all',
-		label: t(ResourceRequests.TABS_ALL),
-		Component: ResourcesAllTab,
-	},
-	{
-		id: 'requests',
-		label: t(ResourceRequests.TABS_REQUESTS),
-		Component: ResourcesRequestsTab,
-	},
-];
-
-const ResourcesAdminPage = () => {
-	const { t } = useTranslation(i18Namespace.resources);
-	const tabs = useMemo(() => createAdminResourcesTabs(t), [t]);
+const ResourcesPage = () => {
+	const tabs = useGetResourcesPageTabs();
 	const { activeTab, setActiveTab } = useTabs(tabs);
 
 	const ActiveComponent = activeTab.Component;
@@ -44,4 +17,4 @@ const ResourcesAdminPage = () => {
 	);
 };
 
-export default ResourcesAdminPage;
+export default ResourcesPage;
