+++ src/pages/admin/ResourcesPage/ui/tabs/ResourcesAllTab/ResourcesAllTab/ResourcesAllTab.module.css
