+++ src/pages/admin/ResourcesPage/ui/tabs/ResourcesAllTab/ResourcesAllTabPagination/ResourcesAllTabPagination.module.css
