--- src/pages/admin/ResourcesPage/ui/tabs/ResourcesAllTab/index.ts
@@ -1,3 +1,0 @@
-export { ResourcesAllTabAsync } from './ui/ResourcesAllTab/ResourcesAllTab.lazy';
-export type { ResourcesAllTabState } from './model/types/resourcesAllTabTypes';
-export { resourcesAllTabReducer } from './model/slice/resourcesAllTabSlice';
