--- src/pages/admin/ResourcesPage/ui/tabs/ResourcesAllTab/ui/ResourcesAllTab/ResourcesAllTab.lazy.tsx
@@ -1,3 +1,0 @@
-import { lazy } from 'react';
-
-export const ResourcesAllTabAsync = lazy(() => import('./ResourcesAllTab'));
