--- src/pages/admin/ResourcesPage/ui/tabs/ResourcesRequestsTab/index.ts
@@ -1,4 +1,0 @@
-export { ResourcesRequestsTab } from './ui/ResourcesRequestsTab/ResourcesRequestsTab';
-export type { ResourcesRequestsTabState } from './model/types/resourcesRequestsTabTypes';
-export { resourcesRequestsTabReducer } from './model/slice/resourcesRequestsTabSlice';
-export { ResourcesRequestsTabPagination } from './ui/ResourcesRequestsTabPagination/ResourcesRequestsTabPagination';
